// web/lib/api/generated/model/notificationDeliveriesStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type NotificationDeliveriesStatus = typeof NotificationDeliveriesStatus[keyof typeof NotificationDeliveriesStatus];


export const NotificationDeliveriesStatus = {
  FAILED: 'FAILED',
  PENDING: 'PENDING',
  SENDING: 'SENDING',
  SENT: 'SENT',
  SKIPPED: 'SKIPPED',
} as const;